+++ lib/Postfilter/Codes.pm
@@ -135,6 +135,8 @@ my %MESSAGE = (
     112 => 'Probable Base64 attachment is forbidden in text groups',
     113 => 'Malformed MIME structure in text article',
     114 => 'Unapproved multipart container is forbidden in text groups',
+    115 => 'Invalid Newsgroups syntax',
+    116 => 'Invalid Followup-To syntax',
 );
 
 my %SPECIAL_SYMBOLIC = (
@@ -262,6 +264,8 @@ my %CATEGORY = (
     112 => 'BODY',
     113 => 'MIME',
     114 => 'MIME',
+    115 => 'GROUP',
+    116 => 'GROUP',
 );